requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
Django==4.2
django-bootstrap5==24.3
django-extensions==3.2.0
frozenlist==1.7.0
geoip2==5.1.0
idna==3.10
maxminddb==2.8.2
multidict==6.6.4
pillow==11.3.0
propcache==0.3.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.3
requests==2.32.5
soupsieve==2.7
sqlparse==0.5.0
typing_extensions==4.14.1
tzdata==2024.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
yarl==1.20.1
